+++ src/sfp/__init__.py
@@ -1,2 +1,2 @@
-from ._poller import Poller, Parameters, dummy_file_check, dummy_file_processing, simple_logging
+from ._poller import Poller, Parameters, dummy_file_check, dummy_file_processing, dummy_batch_processing, simple_logging
 from ._poller import GLOB_NAME_PLACEHOLDER, LOGGING_TYPE_DEBUG, LOGGING_TYPE_INFO, LOGGING_TYPE_ERROR